feat : 게임 기능 마무리
- 경험치, 레벨업 시스템 추가
- 아이템 드롭 추가
- 스토리 선택 추가
- 몬스터 종류별 무기, 갑옷 변경

+++ java/src/main/java/com/byulstudy/utils/RandomUtils.java
@@ -9,12 +9,12 @@ public class RandomUtils {
         return RANDOM.nextInt(inclusiveBegin, exclusiveEnd);
     }
 
-    public static boolean getBooleanByPercentage(final int percentage) {
+    public static boolean getBooleanByPercentage(final double percentage) {
         validatePercentage(percentage);
-        return Math.random() < (double) percentage / 100;
+        return Math.random() <  percentage / 100;
     }
 
-    private static void validatePercentage(final int percentage) {
+    private static void validatePercentage(final double percentage) {
         if (percentage < 0 || percentage > 100) {
             throw new IllegalArgumentException("percentage must be less than 100 and greater than 0");
         }
